cloudFunctions/requirements.txt
cloudsql==0.1.0
Flask==2.3.1
frozenlist==1.3.3
functions-framework==3.3.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
Jinja2==3.1.2
mysql-connector==2.2.9
requests==2.28.2
